add middleware for jwt token authentication from request
